sending ether updated
