+++ example_code/basic_examples/sending_ether/Cargo.toml
@@ -1,9 +1,12 @@
 [package]
-name = "stylus-sending-ether"
+name = "stylus_sending_ether_example"
 version = "0.1.7"
 edition = "2021"
+license = "MIT OR Apache-2.0"
+homepage = "https://github.com/OffchainLabs/stylus-hello-world"
+repository = "https://github.com/OffchainLabs/stylus-hello-world"
 keywords = ["arbitrum", "ethereum", "stylus", "alloy"]
-description = "Stylus sending ether example"
+description = "Stylus hello world example"
 
 [dependencies]
 alloy-primitives = "0.7.6"
@@ -19,9 +22,10 @@ eyre = "0.6.8"
 
 [features]
 export-abi = ["stylus-sdk/export-abi"]
+debug = ["stylus-sdk/debug"]
 
 [[bin]]
-name = "stylus-multi-call"
+name = "stylus-hello-world"
 path = "src/main.rs"
 
 [lib]
